Casi terminado el back

+++ src/main/java/org/atos/dual/club_nautico_api/Repository/BarcoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BarcoRepository extends JpaRepository<Barco, Integer> {
+public interface BarcoRepository extends JpaRepository<Barco, Long> {
 
 }
