+++ src/main/java/org/atos/dual/club_nautico_api/Repository/MiembroRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MiembroRepository extends JpaRepository<Miembro, Integer> {
+public interface MiembroRepository extends JpaRepository<Miembro, Long> {
 
 }
