+++ src/main/java/org/atos/dual/club_nautico_api/Repository/ViajeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ViajeRepository extends JpaRepository<Viaje, Integer> {
+public interface ViajeRepository extends JpaRepository<Viaje, Long> {
 }
